fix: move ActiveSessionLandingScreen into the ActiveSessionLogin folder, and delete logics related to spid login
