+++ ts/features/authentication/common/navigation/params/AuthenticationParamsList.ts
@@ -16,7 +16,6 @@ export type AuthenticationParamsList = {
   [AUTHENTICATION_ROUTES.OPT_IN]: ChosenIdentifier;
   [AUTHENTICATION_ROUTES.IDP_SELECTION]: undefined;
   [AUTHENTICATION_ROUTES.IDP_LOGIN]: undefined;
-  [AUTHENTICATION_ROUTES.IDP_ACTIVE_SESSION_LOGIN]: undefined;
   [AUTHENTICATION_ROUTES.AUTH_SESSION]: undefined;
   [AUTHENTICATION_ROUTES.IDP_TEST]: undefined;
   [AUTHENTICATION_ROUTES.CIE_ACTIVATE_NFC_SCREEN]: CieCardReaderScreenNavigationParams;
